backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
mysqlclient==2.2.7
redis==5.0.8
python-dotenv==1.0.1
pytest==8.3.4
pytest-django==4.10.0
drf-spectacular==0.28.0
